split touch methods, and fix swiftlint errors
